Defer statement preparation until it is executed the first time.
This way a statement can be created in a packages init() function which might be executed before the "Default" schema has its defaults set. With this change as long as the Default schema is setup before the first time the statement is executed, those default values will be honoured in the statement.