wip package private/statement
Key goal is to create prepared statements that can be cached independently of the dialect and column naming convention that is used.

Using the name of the dialect and naming convention as keys to the statement cache is a bit limiting, and reduces the flexibility of naming conventions in particular.
This eliminates the need for dialects and naming conventions to be named, as well as moving a the complexity of statements into its own package.