Update checkSQL to reject `delete from tablename`
To avoid any accidental delete of entire table: given that we allow `insert into tablename` and `update tablename` as abbreviations, it might be natural to try `delete from tablename`, which is actually valid SQL, and deletes everything from the table. The easiest thing to do here is just never accept that command. If someone wants to delete everything from a table it is easy to do directly with "database/sql".